@@ -8,25 +8,25 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^8.1",
+        "php": "^8.2",
         "guzzlehttp/guzzle": "^7.2",
-        "laravel/framework": "^10.10",
-        "laravel/sanctum": "^3.2",
-        "laravel/telescope": "^4.16",
+        "laravel/framework": "^11.0",
+        "laravel/sanctum": "^4.0",
+        "laravel/telescope": "^5.0",
         "laravel/tinker": "^2.8",
-        "mcamara/laravel-localization": "1.8",
+        "mcamara/laravel-localization": "^2.0",
         "moonshine/moonshine": "^1.61",
-        "tightenco/ziggy": "^1.6"
+        "tightenco/ziggy": "^2.0"
     },
     "require-dev": {
         "deployer/deployer": "^7.3",
         "itsgoingd/clockwork": "^5.1",
         "fakerphp/faker": "^1.9.1",
-        "laravel/breeze": "^1.21",
+        "laravel/breeze": "^2.0",
         "laravel/pint": "^1.0",
         "laravel/sail": "^1.18",
         "mockery/mockery": "^1.4.4",
-        "nunomaduro/collision": "^7.0",
+        "nunomaduro/collision": "^8.1",
         "phpunit/phpunit": "^10.1",
         "grkamil/laravel-telegram-logging": "^1.10",
         "spatie/laravel-ignition": "^2.0"
